// src/types/DefaultData.ts
export interface DefaultData {
    name: string
    login: string
    avatar_url?: string
    url: string
    bio?: string
    created_at: string
    followers: number
    following: number
    public_repos: number
    location?: string
    email?: null
    blog?: string
    twitter_username?: null
    company?: string
}